+++ evals/acs/scenarios/__init__.py
@@ -9,6 +9,7 @@ from . import s02_create_ticket_artifacts
 from . import s03_resume_and_verify
 from . import s04_skill_triggers
 from . import s05_session_end
+from . import s06_update_migration
 
 SCENARIOS = [
     s01_install_gate_smoke,
@@ -16,4 +17,5 @@ SCENARIOS = [
     s03_resume_and_verify,
     s04_skill_triggers,
     s05_session_end,
+    s06_update_migration,
 ]